+++ changelog.md
@@ -1,3 +1,9 @@
+# 1.0.1
+- Added Custom Move Buttons (1/2, 1/4, 1/8, Unit).
+- Added Set Current Group button.
+- Added Select Filter toggler.
+- Added Start Flipped toggler.
+
 # 1.0.0
 - Added ability to color objects to 3DL and White.
 - Added Go To All and Next Free Layer buttons.
@@ -6,3 +12,4 @@
 - Added Go To Group button.
 - Added Free Scroll (temporary).
 - Added Copy Object String.
+- Added Select Filter.